Refactor src and deployWorker
